feat(rds): Add log group property to RDS instance and RDS cluster (#28676)
There isn't an easy way to get the new created log group of an DatabaseInstance if the DatabaseInstance has the cloudwatchLogsExport property set. Right now users would have to manual figure out the log group name by constructing `/aws/rds/instance/${this.databaseInstance.instanceIdentifier}/postgresql`. Same applies to RDS cluster resource.

This is also dangerous as it relies on the naming convention unmodified. It would break the application if the naming convention was somehow changed.

Add a new property `cloudwatchLogGroups` to RDS instance and cluster.

Closes #<https://github.com/aws/aws-cdk/issues/20358>.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*