feat(efs): One-Zone filesystem (#28501)
In this PR, I have added support for creating one zone filesystems. 

[As discussed in the issue](https://github.com/aws/aws-cdk/issues/15864#issuecomment-895483167), I considered including `availabilityZoneName` in the props, but for now, I have added only the `oneZone` argument.

The `efs.FileSystem` construct also generates a mount target simultaneously. However, when creating a one zone file system, both the mount target and the file system should be located in the same Availability Zone. The issue arises because the name of the Availability Zone where the mount target will be placed (obtained from `props.vpc.selectSubnets()`) remains token, making it impossible to compare it with the given `availabilityZoneName`.

Please let me know if there is a smart way to resolve this.

Closes #15864

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*