fix(appsync): add dependency to logretention for graphql apis log group (#28548)
This change adds a dependency to log retention if logconfig is set for `GraphqlApi`. This in turn avoids the race condition when subsequent resources refer to the log group property of `GraphqlApi.logGroup`

logGroup is now referring to the custom resource's output log ARN and further resources will depend on the custom resource to be created first.

Closes #26564

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*