fix: deployment recreated on every cdk deployment
Every time I perform a CDK deployment, the logicalId hash of the
deployment resource changes and causes the deployment resource to be
deleted and recreated. I'm assuming it is because the configuration
`content` is part of the hash creation and I am creating the content
using `lazy` and at the time of the hash creation it is still a token.

Looking at the [CloudFormation
docs](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-appconfig-deployment.html)
for the deployment resource, a change to _any_ property causes a
replacement so I don't think we need to control the recreation logic
ourselves, we should just let CloudFormation do the resource replacement
for us.