fix(custom-resources): custom resource response may silently get dropped
It looks like in our custom resource response senders, we are not
checking the HTTP result code of making the `PUT` request to the
presigned URL.

This allows for the possibility that we don't detect that we got a
500, exit successfully, and drop the response on the floor.

Also log the URL we're writing to so that it becomes easier to debug
missing object PUTs.

Yes, I am aware that there is awkward duplication going on here. I'm
optimizing for quick delivery here.