fix(eks): helm uninstall in custom resource handler does not respect `Wait` (#28830)
I have a stack that installs and uninstalls helm charts, where the underlying resources use [Kubernetes Finalizers](https://kubernetes.io/docs/concepts/overview/working-with-objects/finalizers/). CDK's helm construct currently ignores `--wait`, which means that any object finalization in those helm charts is not respected in my dependency ordering.

## Compatibility

I could see some debate around whether or not this is a breaking change. I'm currently viewing it as a bug fix. I'm a bit cautious about how valuable it would be to gate this behavior compared to the additional complexity.

## Testing

Tested on a local fork and it worked like a charm

Closes #28831

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*