chore: reformat pull request template (#28923)
### Reason for this change

I noticed that some PRs only link the issue or have really long paragraph descriptions which can be hard to follow and understand the why/what regarding the changes.

### Description of changes

Reformatting the pull request template with section headings. This enforces PR descriptions to follow a uniform format and include the information we want. Should make it easier/more efficient to review PRs. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*