feat: allow overriding the filename of the processed file after substitution (#29029)
### Issue # (if applicable)

Closes #28450 

### Reason for this change

Currently, the filename is automatically generated based on its fingerprint. There's no way to override the resulting filename.

### Description of changes

Added an optional `destinationKey` parameter. If this is provided, the construct would use the `destinationKey` when uploading the resulting processed file to S3. If the `destinationKey` is not provided, the current approach would be the default.

### Description of how you validated changes

Added a unit test to confirm that the correct filename is being used.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*