fix(stepfunctions): support json path fields for item batcher
The ItemBatcher property of DistributedMap supports json path fields,
but the new DistributedMap construct doesn't call
FieldUtils.renderObject when rendering the ItemBatcher.

Make sure to call `FieldUtils.renderObject()` when rendering this
property.

Added an integration test to ensure that the json path field resolves to
the actual value from the state in the input by checking the output from
the results writer in S3.