fix(stepfunctions): escaped curly braces result in error during state machine execution (#29267)
### Issue # (if applicable)

Closes #29261

### Reason for this change

Enable proper escaping of curly braces in Step Functions intrinsic functions. The user needs to be able to explicitly escape curly braces to distinguish them from curly braces used for intrinsic function arguments (`\{\}` vs `{}`.  Currently, the escaping logic double-escapes an already-escaped curly brace. The state machine successfully deploys, but then fails to execute.

### Description of changes

When a backslash immediately precedes a curly brace, it will not be escaped.

### Description of how you validated changes

Added a case to an existing integration test, and updated the integ test to execute the deployed state machine to check for runtime errors.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*