docs(events): mention property requirements when creating custom events (#29287)
### Issue # (if applicable)

Closes #<issue number here>.

### Reason for this change

When creating an EventPattern for a custom event, a few properties are required. Documentation currently isn't clear on which properties are required. This change links to AWS documentation that describes the required properties for custom events.

### Description of changes

Added text stating that some optional properties are required for custom events and provided link to docs. 

### Description of how you validated changes

No validation

### Checklist
- [ ] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*